[indic] Move shared indic data to a separate file
