+++ src/opentype/shapers/indic-data.js
@@ -1,0 +1,35 @@
+// Visual positions in a syllable from left to right.
+export const POSITIONS = {
+  Start: 0,
+
+  Ra_To_Become_Reph: 1,
+  Pre_M: 2,
+  Pre_C: 3,
+
+  Base_C: 4,
+  After_Main: 5,
+
+  Above_C: 6,
+
+  Before_Sub: 7,
+  Below_C: 8,
+  After_Sub: 9,
+
+  Before_Post: 10,
+  Post_C: 11,
+  After_Post: 12,
+
+  Final_C: 13,
+  SMVD: 14,
+
+  End: 15
+};
+
+export const IS_CONSONANT = {
+  C: true,
+  Ra: true,
+  CM: true,
+  V: true,
+  Placeholder: true,
+  Dotted_Circle: true
+};
